backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
opencv-python
python-dotenv
